requirements.txt
aixplain>=0.2.27
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
sqlite3
feedparser>=6.0.0
click>=8.1.0
tqdm>=4.67.1
flask>=2.3.0
flask-cors>=4.0.0